shift RealLocationWebRepository to a new file
